Extracted modals from practitioner and parent calendars and added smoke tests
