+++ speedy-recovery-client/src/__tests__/components/CalendarPages.test.js
@@ -2,9 +2,9 @@ import React from "react";
 import Enzyme, { mount, shallow } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import exampleAppointment from "../test_input/internal/ExampleAppointment";
-import ParentCalendar from "../../components/secured/calendar/ParentCalendar";
-import PatientCalendar from "../../components/secured/calendar/PatientCalendar";
-import PractitionerCalendar from "../../components/secured/calendar/PractitionerCalendar";
+import ParentCalendar from "../../components/secured/calendar/parent/ParentCalendar";
+import PatientCalendar from "../../components/secured/calendar/patient/PatientCalendar";
+import PractitionerCalendar from "../../components/secured/calendar/practitioner/PractitionerCalendar";
 import exampleAppointments from "../test_input/internal/ExampleAppointments";
 import BigCalendar from "react-big-calendar";
 import moment from "moment";
