+++ speedy-recovery-client/src/__tests__/test_input/internal/ExampleAppointment.json
@@ -5,7 +5,6 @@
   "reason": "General Discussion",
   "priority": 5,
   "description": "test",
-  "start": "test",
   "end": "test",
   "comment": "test",
   "patient": "test",
