+++ speedy-recovery-client/src/components/secured/calendar/patient/PatientCalendar.js
@@ -2,7 +2,7 @@ import React from "react";
 import BigCalendar from "react-big-calendar";
 import "react-big-calendar/lib/css/react-big-calendar.css";
 import "react-big-calendar/lib/addons/dragAndDrop/styles.less";
-import "./CalendarPages.css";
+import "../CalendarPages.css";
 import { Grid, Segment } from "semantic-ui-react";
 
 class PatientCalendar extends React.Component {
