+++ speedy-recovery-client/src/service/FhirDataQueryingService.js
@@ -27,9 +27,12 @@ const getUserAppointments = userID => {
 };
 
 // TODO: Also add error handling and return [] if (!appointments.data || !appointments.data.total)
-const extractResourcesFromBundle = appointments => 
-    appointments.data.total !== 0 ? appointments.data.entry.map(app => app.resource) : [];
+const extractResourcesFromBundle = appointments =>
+  appointments.data.total !== 0
+    ? appointments.data.entry.map(app => app.resource)
+    : [];
 
 export default {
-  getUserAppointments, extractResourcesFromBundle
+  getUserAppointments,
+  extractResourcesFromBundle
 };
